Dbg | error pass
