+++ sal-core-macros/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "logging"
+name = "sal-core-macros"
 version = "0.1.0"
 edition = "2024"
